refactor(www): adapt last imports and delete withDefaultSSR implementation
