+++ apps/www/pages/vote/genossenschaft/diskutieren.js
@@ -2,7 +2,7 @@ import ElectionDiscussionPage from '../../../components/Vote/Legacy/ElectionDisc
 import compose from 'lodash/flowRight'
 import { enforceMembership } from '../../../components/Auth/withMembership'
 import withMe from '../../../lib/apollo/withMe'
-import withDefaultSSR from '../../../lib/hocs/withDefaultSSR'
+import { withDefaultSSR } from '../../../lib/apollo/helpers'
 
 export default withDefaultSSR(
   compose(enforceMembership(), withMe)(ElectionDiscussionPage),
